+++ docs/public_api.opml
@@ -348,6 +348,9 @@
           <outline text="func|pixeltable.functions.vision.bboxes_resize_canvas" />
           <outline text="func|pixeltable.functions.vision.overlay_segmentation" />
         </outline>
+        <outline text="class|pixeltable.functions.CodecContextMetadata" />
+        <outline text="class|pixeltable.functions.ContainerMetadata" />
+        <outline text="class|pixeltable.functions.StreamMetadata" />
       </outline>
 
       <!-- AI Integrations group -->
