+++ pixeltable/functions/audio.py
@@ -11,17 +11,19 @@ import av
 import numpy as np
 
 import pixeltable as pxt
-import pixeltable.utils.av as av_utils
 from pixeltable import exceptions as excs
 from pixeltable.env import Env
+from pixeltable.utils import av as av_utils
 from pixeltable.utils.code import local_public_names
 from pixeltable.utils.local_store import TempStore
+
+from . import util
 
 _logger = logging.getLogger('pixeltable')
 
 
 @pxt.udf(is_method=True)
-def get_metadata(audio: pxt.Audio) -> av_utils.ContainerMetadata:
+def get_metadata(audio: pxt.Audio) -> util.ContainerMetadata:
     """
     Gets various metadata associated with an audio file and returns it as a dictionary.
 
@@ -61,7 +63,7 @@ def get_metadata(audio: pxt.Audio) -> av_utils.ContainerMetadata:
 
         >>> tbl.select(tbl.audio_col.get_metadata()).collect()
     """
-    return av_utils.get_metadata(audio)
+    return util.get_metadata(audio)
 
 
 @pxt.udf()
